requirements.txt
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-gitlab>=3.0.0
requests>=2.28.0
gunicorn>=21.0.0
whitenoise>=6.4.0
urllib3